+++ lib/cldr/language_tag.ex
@@ -127,7 +127,7 @@ defmodule Cldr.LanguageTag do
   nu        | Number system identifier         | arabext, armnlow, roman, tamldec
   rg        | Region override                  | The value is a unicode_region_subtag for a regular region (not a macroregion), suffixed by "ZZZZ"
   sd        | Subdivision identifier           | A unicode_subdivision_id, which is a unicode_region_subtagconcatenated with a unicode_subdivision_suffix.
-  ss        | Break suppressions identifier     | none, standard
+  ss        | Break suppressions identifier    | none, standard
   tz        | Timezone identifier              | Short identifiers defined in terms of a TZ time zone database
   va        | Common variant type              | POSIX style locale variant